replace ApiGateway with AppSync and replace handlers for graphql schema and vtl files
