+++ ApiGatewayLambdaHandlers/add-list-item.js
@@ -1,13 +1,14 @@
 /*
-ROUTE : PATCH /list-item  
+ROUTE : POST /list-item  
 */
 
 
 const AWS = require('aws-sdk');
 AWS.config.update({ region: 'us-west-2' })
 
-const util = require('./util.js');
+const util = require('../api/util.js');
 const moment = require('moment');
+const { v4: uuidv4 } = require('uuid');
 
 const dynamoDB = new AWS.DynamoDB.DocumentClient();
 const tableName = process.env.LIST_ITEM_TABLE;
@@ -17,24 +18,19 @@ exports.handler = async (event) => {
         let item = JSON.parse(event.body).Item;
         item.user_id = util.getUserId(event.headers);
         item.user_name = util.getUserName(event.headers);
+        item.list_item_id = item.user_id + ':' + uuidv4();
+        item.timestamp = moment().unix();
         item.expires = moment().add(60, 'days').unix();
 
-       await dynamoDB.put({
+        await dynamoDB.put({
             TableName: tableName,
-            Item: item,
-            ConditionExpression: '#t = :t',
-            ExpressionAttributeNames: {
-                '#t': 'timestamp'
-            },
-            ExpressionAttributeValues: {
-                ':t': item.timestamp
-            }
+            Item: item
         }).promise();
 
         return {
             statusCode: 200,
             headers: util.getResponseHeaders(),
-            body: JSON.stringify({item})
+            body: JSON.stringify(item)
         }
     } catch (error) {
         console.log("Error", error)
