+++ api/mapping-templates/requests/create-list-item.vtl
@@ -1,0 +1,8 @@
+{
+    "version": "2017-02-28",
+    "operation": "Invoke",
+    "payload": {
+        "field": "createListItem",
+        "arguments":  $utils.toJson($context.arguments)
+    }
+}
