+++ api/mapping-templates/requests/get-list-item.vtl
@@ -1,0 +1,8 @@
+{
+    "version": "2017-02-28",
+    "operation": "Invoke",
+    "payload": {
+        "field": "getListItem",
+        "arguments":  $utils.toJson($context.arguments)
+    }
+}
