+++ api/mapping-templates/requests/get-list-items.vtl
@@ -1,0 +1,8 @@
+{
+    "version": "2017-02-28",
+    "operation": "Invoke",
+    "payload": {
+        "field": "getListItems",
+        "arguments":  $utils.toJson($context.arguments)
+    }
+}
