+++ api/mapping-templates/requests/update-list-item.vtl
@@ -1,0 +1,8 @@
+{
+    "version": "2017-02-28",
+    "operation": "Invoke",
+    "payload": {
+        "field": "updateListItem",
+        "arguments":  $utils.toJson($context.arguments)
+    }
+}
