+++ api/mapping-templates/responses/generic-response.vtl
@@ -1,0 +1,1 @@
+$util.toJson($ctx.result)
