+++ package.json
@@ -7,7 +7,10 @@
     "aws-sdk": "^2.710.0",
     "moment": "^2.27.0",
     "serverless-appsync-plugin": "^1.3.1",
+    "serverless-iam-roles-per-function": "^2.0.2",
     "serverless-offline": "^6.4.0",
+    "serverless-plugin-additional-stacks": "^1.4.0",
+    "serverless-plugin-include-dependencies": "^4.1.0",
     "underscore": "^1.10.2",
     "uuid": "^8.2.0"
   },
