+++ resources/dyanamo-table.yml
@@ -1,0 +1,61 @@
+ListItemTable:
+Type: AWS::DynamoDB::Table
+DeletionPolicy: Retain
+Properties:
+  RoleName: ${self:custom.role.dynamodb.name}
+  AssumeRolePolicyDocument:
+    Version: "2012-10-17"
+    Statement:
+      - 
+        Effect: "Allow"
+        Principal:
+          Service:
+            - "appsync-amazonaws.com"
+        Action
+          - "sts:AssumeRole"
+  Policies:
+    -
+    PolicyName: ${self:custom.role.dynamodb.name}-Policy
+    PolicyDocument:
+      Version: "2012-10-17"
+      Statement:
+        - 
+          Effect: "Allow"
+          Action:
+            - "dynamodb:Query"
+            - "dynamodb:BatchWriteItem"
+            - "dynamodb:GetItem"
+            - "dynamodb:DeleteItem"
+            - "dynamodb:PutItem"
+            - "dynamodb:UpdateItem"
+            - "dynamodb:BatchGetItem"
+          Resource:
+            - "arn:aws:dynamodb:${opt:region, self:provider.region}:*:table/${self:provider.environment.LIST_ITEM_TABLE"
+            - "arn:aws:dynamodb:${opt:region, self:provider.region}:*:table/${self:provider.environment.LIST_ITEM_TABLE/*"
+    
+
+  AttributeDefinitions:
+    - AttributeName: user_id
+      AttributeType: S
+    - AttributeName: timestamp
+      AttributeType: N
+    - AttributeName: list_item_id
+      AttributeType: S
+  KeySchema:
+    - AttributeName: user_id
+      KeyType: "HASH"
+    - AttributeName: timestamp
+      KeyType: "RANGE"
+  ProvisionedThroughput:
+    ReadCapacityUnits: 1
+    WriteCapacityUnits: 1
+  GlobalSecondaryIndexes:
+    - IndexName: list_item_id-index
+      KeySchema:
+        - AttributeName: list_item_id
+          KeyType: "HASH"
+      Projection:
+        ProjectionType: ALL
+      ProvisionedThroughput:
+        ReadCapacityUnits: 1
+        WriteCapacityUnits: 1
